Remove unnecessary device select pin manipulation (#137)
Resolves #136 (Remove unnecessary device select pin manipulation).

The SPI peripheral based SPI controller echo interactive test previously
configured the device select pin as a push-pull I/O pin before the SPI
controller was constructed and initialized. This did not reflect the
order in which these operations would be performed in an actual
application (the SPI controller would be initialized before the devices
and their associated device selectors), and was not necessary since the
device select pin should be weakly pulled up if the SPI peripheral is
being used as a controller.